LayoutTests/fast/viewport/viewport-12.html

<head>
    <title>Viewport META with "maximum-scale=0.5, initial-scale=4".</title>
    <meta name="viewport" content="maximum-scale=0.5, initial-scale=4">
    <script>
        function test() {
            if (window.testRunner) {
                testRunner.dumpAsText();
                alert(internals.configurationForViewport(document, 1, 230, 480, 320, 352));
            }
        }
    </script>
</head>
<body onload="test();">
