safari-600.7-branch/LayoutTests/fast/viewport/viewport-12-expected.txt

ALERT: viewport size 640x704 scale 0.5 with limits [0.5, 0.5] and userScalable true
